Allow workers to get new accounts when banned
This adds an option -w/--workers to specify how many search workers to create.
If unspecified, it defaults to the number of accounts specified, mimicking existing behaviour

Accounts are added to a queue, and wokers pull an account from the queue when needed.
When a worker has an exception or fails too many times, it will put it's account back
on the end of the queue, and grab one from the top of the queue.

This allows workers to continue processing on fresh accounts, and leaves potentially banned
accounts in the queue for retrying later.